+++ typescript-client-sdk/src/sdk/sdk.ts
@@ -114,8 +114,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "0.1.0";
-    sdkVersion = "1.26.0";
-    genVersion = "2.82.0";
+    sdkVersion = "1.27.0";
+    genVersion = "2.84.1";
     globals: any;
 
     public constructor(init?: Partial<SDKConfiguration>) {