+++ RELEASES.md
@@ -355,3 +355,18 @@ Based on:
 - Speakeasy CLI 1.69.1 (2.82.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [Go v1.26.0] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.26.0 - go-client-sdk
+
+## 2023-08-15 00:11:23
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 
+- Speakeasy CLI 1.72.0 (2.84.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.27.0] python-client-sdk
+- [ruby v1.27.0] ruby-client-sdk
+- [typescript v1.27.0] typescript-client-sdk
+- [go v1.27.0] go-client-sdk
+- [java v1.27.0] java-client-sdk
+- [php v1.27.0] php-client-sdk
+### Releases
+- [Go v1.27.0] https://github.com/speakeasy-api/openapi-generation-tests/releases/tag/go-client-sdk/v1.27.0 - go-client-sdk